On load of main, dataGridView Data loads and displays data from the movie table.
In Main.cs, I added the void method loadMovieGrid(), and in ConnectionClass.cs, I added the method loadMovieData(), which returns a DataTable with the data from the Movie table.

In Moive.cs, I had to change the "Movie" class to "Movie1" because PullMovieAPI already had a "Movie" class.  I made this change so I could build and test the solution.

Let me know if there is any problems with my code.